Change stream management for insertion and pick up.
Statuses used to be fetched and inserted only in `statuses` table - and
only when picking one up for the stream they were added to
`streamcontent`.
Now, statuses fetched from Mastodon are immediately added to the
streamcontent, but without a given position.

Eventually, that will allow to have the `statuses` table only act as a
cache - making it possible for example to drop the actual content if
needed.

It also makes the data a bit more logical, where the potential content
of a stream is all in a table.