Start splitting the main rendering component.
That is in preparation of having a search result page.

It creates a kind of template component to serve as the main UI, where
the concrete component can customize the various part (menu, stream,
footer).

It means that in the future where there is more than one page, it will
change the whole UI instead of just things in place - so the structure
might change, but for now it should do.