Rework UI CSS.
Now the UI elements (header/footer) are an overlay which is absolutely
positioned. It makes a bit more sense (think game UI overlay), but more
importantly it makes behavior when there is no status better. Instead of
having the footer be sometimes in the middle of the page, it is always
at the bottom. It also fixes background color, and allow for some
specific color on the middle column when there is no status.