Reduce jitter of footer when scrolling on mobile.
Scrolling on mobile sometimes make the address bar appear/disappear.
This in turns changes the size of the view port, and require fixed
element to recompute

When using a page-wide element, the resizing comes too late, showing the
footer moving around, before settling down when the scrolling is done.

By removing the UI overlay element, now the footer can be directly
marked as a `bottom: 0` fixed element. This fixes the jumping around on
Chrome, as the reference point seem to be better and account for the
address bar appearing/disappearing animation.

On firefox, I think it improves things, but it is still jittery - the
footer clearly moves a bit around when scrolling & triggering address
bar visibility.