Fiddle with initial loading of statuses.
It used to try to give some already-read status on initial load. The
idea was to give a sense of context and re-assurance that statuses were
not missed. In practice, it was not working well, and using up quite
some space, esp. with long statuses. Moreover, last-read status is just
based on what was visible on screen, while people are likely to have
read to the bottom of their screen.

Now, it will load just starting from the first unread status.

Also, initial loading will stop triaging new statuses. The goal here is
to make "UI loading" as neutral as possible. When automatic loading &
triaging of new statuses is done, it will be different - but for now the
model is explicit action to get more triaged statuses.