Fix notifications.
Getting the list of notifications is not helpful - it is in its own just
a stream from mastodon perspective. What is important is to get what was
the last one which was read - and that is done by something called
"markers". So now, the marker is first read and then notifications are
obtained to see how many are more recents.

Unfortunately, go-mastodon does not support markers, so that requires
having a custom GetMarker call.