Fix case where there is no status on initial load.
If the last-read marker is set to the latest status available, first
initial list when loading the UI is going to be empty. As the code used
to use the position attached to provided statuses to issue following
list requests, it was failing as position was not available.

Instead, use the continuation markers provided by the List API.

This is pretty much untested.