Change indices on streamcontent table.
The index on stid was meaningless, as it has a lot of rows for a single
value. Add an index on position to be able to easily find recent
position - which is the most common operation.

Also, run optimize to make sure the optimizer has a chance to pick the
right index.