Switch back in-memory DBs to split ro/rw connections.
The code was recently changed to allow for nested queries. That was
working in all cases except for the in-memory version, as it was
limiting the Go SQL DB object to only a single connection. The change was
to keep using a single DB, but with multiple SQL connections. But that
in turn broke the case of the in-memory DB, as each connection was
seeing different content, leading to various breakages.

So, this change moves to using `cache=shared` to ensure that all
connections arrive at the same in-memory table. In-turn, that requires
to have a differnt DB object for read-only and read-write - where the
read-write DB only allows for a single connection. This avoids "database
is locked" kind of errors, as all write requests are serialized.

This has also the advantage of making in-memory case more similar to the
case where actual files exist. What I'm not clear is that used to be the
case, but was changed in an older commit:
 https://github.com/Palats/mastopoof/commit/359e78beadb4467b0bc7b994900040138246e49a

And from the description, it means that what I'm (re-)introducing here
should generate the same problems (lock issues?), but I'm definitly not
seeing that so far. Maybe other things have changed, but I'm not clear
what.