Rename "statusstate" to "statusMeta".
This is data which is coming from Mastodon. This is NOT impacted by
Mastopoof stream state, so the "state" part was making things a bit hard
to follow - esp. given I want to add more state to the stream.

This does not update the DB schema yet - this will be done in another
change.