Remove type mismatch.
There are 2 version of `StatusItem` in the frontend. That is a bit
terrible, but a typescript assertion was added to ensure that one was a
subset of the other. Unfortunately, the linter was also deactivated on
that assertion to avoid a message indicating that the asset function
(which was just here to typecheck) is unused. And that disabled also the
type checking.

Re-enabling the type assertion and fixing the field mismatch.