Introduce an npm workspace to avoid duplicated modules in bundle.
When using relative imports in packages.json, common dependencies such
as @bufbuild/protobuf were ending up inlined twice in the vite build
bundle. As this library is setting up a Javascript Symbol() at the top
level, it meant the code was ending up with 2 different versions of that
symbol. As this symbol is used to get metadata about an enum, that lead
to enums not working: one version of the symbol was used when creating
something from proto/, and another version of the symbol was used when
manipulating the data from frontend/.

With NPM workspaces at the top level, it seems the bundling does the
right thing, and notices that both proto/ and frontend/ end up
dependending on a common library - and thus avoid inlining them twice.

That required to also fiddle with the font loading. It was a big hack
before, accessing the file directly in node_modules. But with
node_modules being in the root of the repository, the dev version was
not working anymore, as it was (I think) unhappy with escaping to the
parent directory.

Instead, include the font from an import in Javascript. This delays the
loading of the font after Javascript is loaded, which is not ideal -
that's a problem to fix later, maybe with vite-plugin-webfont-dl.